# openapi.yaml
openapi: 3.0.1
info:
  title: AI-Powered Shopping Assistant
  description: A plugin that allows the user to search for product with AI.
  version: 'v1'
servers:
  - url: http://localhost:5003
paths:
  /google_search/{query}:
    get:
      operationId: get_google_search_results
      summary: Get Google Custom Search results
      description: Retrieves search results from Google Custom Search based on the query parameter.
      parameters:
        - in: path
          name: query
          schema:
            type: string
          required: true
          description: The search query to use when fetching results from Google Custom Search.
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          required: false
          description: The page number to fetch results for, with each page containing up to 5 results.
        - in: query
          name: num
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 5
          required: false
          description: The page number to fetch results for, with each page containing up to 5 results.
        - in: query
          name: start_date
          schema:
            type: string
            format: date
          required: false
          description: The start date for the date range filter in the format YYYYMMDD.
        - in: query
          name: end_date
          schema:
            type: string
            format: date
          required: false
          description: The end date for the date range filter in the format YYYYMMDD.
      responses:
        '200':
          description: Search results retrieved successfully
          content:
            application/json:
              schema:
                type: object
                description: The search results in JSON format.
                properties:
                  assistant:
                    type: string
                    description: Assistant's message about the total results found.
                  Data:
                    type: object
                    description: The raw data retrieved from the Google Custom Search API.
                  assistant_message:
                    type: string
                    description: Assistant's message guiding the user to ask for more results.
                  current_page:
                    type: integer
                    description: The current page number of the search results.
                  total_results:
                    type: integer
                    description: The total number of search results found.
                  results:
                    type: array
                    items:
                      type: object
                      description: An array of search result items.
                      properties:
                        index:
                          type: integer
                          description: The index of the search result item.
                        title:
                          type: string
                          description: The title of the search result item.
                        link:
                          type: string
                          description: The link to the search result item.
                        snippet:
                          type: string
                          description: A snippet of the search result item.
        '500':
          description: Server error, e.g., the server encountered a situation it doesn't know how to handle, such as an unexpected exception.